AWS platform code is now organized in its own subdirectory, which will make it easier to break up the large platform.ts file into smaller, more manageable pieces
