+++ apps/cli/src/platforms/aws/stack-config.ts
@@ -1,9 +1,9 @@
 
 import { CloudFormationClient, DescribeStacksCommand } from '@aws-sdk/client-cloudformation';
-import { StackOutput, AWSError } from '../lib/types.js';
-import { validateAwsResourceName, assertValid } from '../lib/validators.js';
-import { logger } from '../lib/logger.js';
-import { loadEnvironmentConfig, type EnvironmentConfig } from '../platforms/platform-resolver.js';
+import { StackOutput, AWSError } from '../../lib/types.js';
+import { validateAwsResourceName, assertValid } from '../../lib/validators.js';
+import { logger } from '../../lib/logger.js';
+import { loadEnvironmentConfig, type EnvironmentConfig } from '../platform-resolver.js';
 
 export interface SemiontConfig {
   region: string;
